product create ve editde olan problemler hell edildi

+++ Lupum_Yolcu/Controllers/ProductsController.cs
@@ -46,17 +46,20 @@ namespace Lupum_Yolcu.Controllers
             if (ModelState.IsValid)
             {
 
-                var colors = product.Colors.Split(',');
-                foreach (var item in colors)
+                if (product.Colors != null)
                 {
-                    var pcolor = new ProductColor
+                    var colors = product.Colors.Split(',');
+                    foreach (var item in colors)
                     {
-                        ProductId = product.Id,
-                        Name = item
-                    };
+                        var pcolor = new ProductColor
+                        {
+                            ProductId = product.Id,
+                            Name = item
+                        };
 
-                    _context.ProductColors.Add(pcolor);
+                        _context.ProductColors.Add(pcolor);
 
+                    }
                 }
                 _context.Products.Add(product);
                 _context.SaveChanges();
@@ -84,7 +87,6 @@ namespace Lupum_Yolcu.Controllers
         [HttpPost,ValidateAntiForgeryToken]
         public ActionResult Edit(Product product, string Status,ProductNetworkPrice[] Prices)
         {
-          
             product.Status = true;
             if (string.IsNullOrEmpty(Status))
             {
@@ -94,7 +96,8 @@ namespace Lupum_Yolcu.Controllers
             {
                 ModelState.AddModelError("Name", "Bu adla bazada mehsul var");
             }
-            if (Prices != null)
+            if (!ModelState.IsValid) return View(product);
+            if(Prices != null)
             {
                 _context.ProductNetworkPrices.RemoveRange(_context.ProductNetworkPrices.Where(pnp => pnp.ProductId == product.Id));
                 foreach (var item in Prices)
@@ -102,41 +105,38 @@ namespace Lupum_Yolcu.Controllers
                     item.ProductId = product.Id;
                     _context.ProductNetworkPrices.Add(item);
                 }
-                _context.SaveChanges();
             }
-            else
+            if(product.Colors != null)
             {
-                if (ModelState.IsValid)
+                _context.ProductColors.RemoveRange(_context.ProductColors.Where(pnp => pnp.ProductId == product.Id));
+                var colors = product.Colors.Split(',');
+                foreach (var color in colors)
                 {
-                    _context.Entry(product).State = System.Data.Entity.EntityState.Modified;
-                    _context.ProductColors.RemoveRange(_context.ProductColors.Where(pnp => pnp.ProductId == product.Id));
-
-
-
-                    var colors = product.Colors.Split(',');
-                    foreach (var color in colors)
+                    var pcolor = new ProductColor
                     {
-                        var pcolor = new ProductColor
-                        {
-                            ProductId = product.Id,
-                            Name = color
-                        };
+                        ProductId = product.Id,
+                        Name = color
+                    };
 
-                        _context.ProductColors.Add(pcolor);
+                    _context.ProductColors.Add(pcolor);
 
-                    }
-                    _context.SaveChanges();
-                    return RedirectToAction("Index");
                 }
             }
-            ModelState.AddModelError("", "price bos gelir");
-            ViewBag.ErrorMessage = "Email not found or matched";
-            ViewBag.Types = _context.Types.ToList();
-            ViewBag.Networks = _context.Networks.ToList();
-            product.ProductNetworkPrices = _context.ProductNetworkPrices.Where(pi => pi.ProductId == product.Id).ToList();
-
-
-            return View(product);
+            _context.Entry(product).State = System.Data.Entity.EntityState.Modified;
+            _context.SaveChanges();
+            return RedirectToAction("Index");
+            //if (Prices != null)
+            //{
+               
+            //    if (ModelState.IsValid)
+            //    {
+                    
+            //    }
+            //}
+            //ModelState.AddModelError("", "price bos gelir");
+            //ViewBag.ErrorMessage = "Email not found or matched";
+           
+            //return View(product);
         }
         #endregion
